More refactoring for better encapsulation
This commit moves some stateless function to the service_handler class. I've
always wanted to slim a bit the supervisorU/mod.rs file and this commit is my
take on that.
With this refactoring, there is a small public interfacte from the
service_handler module that is used by the supervisor. Everything else is
nicely private to the package.
